[synaptic] Swipe action based on the three-fingers touch
When touchpad can report more then 2 active fingers, we can generate action
based on swipe gesture. For now it includes swipes to the left and to the
right. One button click event can be assigned to those actions, which by
default (internal driver default) is button 0 - swipes are disabled.

The most preferable button assignment is button 8 and 9 for to the left and
to the right swipes respectively. Such a configuration provides browser
history navigation (e.g. Firefox) with horizontal swipes.

Added Synaptic driver options:
* SwipeLeftButton
* SwipeRightButton
* SwipeThreshold

Signed-off-by: Arkadiusz Bokowy <arkadiusz.bokowy@gmail.com>